Remove stats_screen remnants
